Add files via upload
Setting the right price for a used vehicle can be challenging for individuals with limited car knowledge and the abundance of features that affect pricing. Often, sellers miss out on potential profits by settling for lower offers to expedite the sale. To combat this business problem, Team 10 has developed a solution through a comprehensive analysis of the key attributes influencing vehicle pricing in the United States and the United Kingdom. The storyline of our Python notebook revolves around encouraging sellers to price their used cars for independent sales effectively. To explore this, we analyzed used car market trends, explored car conditions, dove into car specifications, considered the influence of sustainability on consumer choices, and created a machine learning model to predict the price. This cohesive approach not only equips sellers with valuable insights but also enables them to make informed pricing decisions that maximize their profits in a competitive marketplace.